Added PyVkabat.py to repository.
Added the PyVkabat.py file to the repository. This version has 'hard-coded' configuration area which will need to be removed. This submits to various secondary structure prediction algos in parallel to speed up previous versions of the code. This version is heavily inspired by the work of BNS and others in the Dr. Brooke Lustig Lab at SJSU. That version was written in Java and required a local installation of perl and the jpredapi. This version is written exclusively in Python3, does not require the jpredapi (although, as an option, a python version of the jpredapi can be called), does not require a perl installation (although, if the use of the python jpredapi is required so is perl), the code is also parallelized to speed up the calculation.